% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_vaccine_table_format.R
\name{align_bold}
\alias{align_bold}
\title{Load, clean, and analyze state hospital data}
\usage{
align_bold(dataset)
}
\arguments{
\item{dataset}{}
}
\value{
returns aggregate hospital table with metrics
}
\description{
Load, clean, and analyze state hospital data
}
