% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clean_raw_data.R
\name{clean_raw_hosp_nat_data}
\alias{clean_raw_hosp_nat_data}
\title{clean_raw_hosp data}
\usage{
clean_raw_hosp_nat_data(data, date_column = Date, lag = 0, start_date = NULL)
}
\arguments{
\item{data}{historical data table}

\item{date_column}{name of date column to be renamed}

\item{lag}{number of days to remove from the most recent date}

\item{start_date}{first day date to filter}
}
\value{
data table with most recent data for each jurisdiction with historical data nested
}
\description{
Takes historical data and compresses it into values for the most recent day and calculates aggregate/daily values metrics - totals, means, new
}
