% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_table.R
\name{create_summary_table}
\alias{create_summary_table}
\title{create_summary_table}
\usage{
create_summary_table(...)
}
\arguments{
\item{...}{single row data sets with most recent metrics}
}
\value{
summary table
}
\description{
create_summary_table
}
