% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_numbers.R
\name{daily_lab_numbers_exsum}
\alias{daily_lab_numbers_exsum}
\title{Load, clean, and analyze state hospital data}
\usage{
daily_lab_numbers_exsum(dataset)
}
\arguments{
\item{dataset}{}
}
\value{
returns aggregate hospital table with metrics
}
\description{
Load, clean, and analyze state hospital data
}
