% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_metrics.R
\name{get_14day_cum}
\alias{get_14day_cum}
\title{Get rolling cumulative metrics}
\usage{
get_14day_cum(column, days = 14)
}
\arguments{
\item{days}{date range, number of days to calculate sum, default is 14}

\item{data}{column to calculate sum}
}
\value{
sum column
}
\description{
Get rolling cumulative metrics
}
