% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_7day_cum.R
\name{get_7day_cum}
\alias{get_7day_cum}
\title{Get rolling cumulative metrics}
\usage{
get_7day_cum(column, days = 7)
}
\arguments{
\item{days}{date range, number of days to calculate sum, default is 7}

\item{data}{column to calculate sum}
}
\value{
sum column
}
\description{
Get rolling cumulative metrics
}
