% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_7day_cum.R
\name{get_7day_cum_per_100k}
\alias{get_7day_cum_per_100k}
\title{get_7day_cum_per_100k}
\usage{
get_7day_cum_per_100k(column, days = 7)
}
\arguments{
\item{days}{date range, number of days to calculate sum}

\item{data}{column to calculate sum}
}
\value{
sum column
}
\description{
get_7day_cum_per_100k
}
