% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_case_death_tbl.R
\name{get_case_death_juris_tbl}
\alias{get_case_death_juris_tbl}
\title{Load, clean, and analyze case and death tab}
\usage{
get_case_death_juris_tbl(
  username,
  date_column = "submission_date",
  state_abbr = "state_abbr",
  lag = 0,
  start_date = "2020-01-01"
)
}
\arguments{
\item{username}{cpu username of user compiling the report.}

\item{date_column}{name of date column to be renamed}

\item{state_abbr}{name of column with state and territory abbreviations to be renamed and standardize abbreviations}

\item{lag}{number of days to remove from the most recent date}

\item{start_date}{first day date to filter}
}
\value{
returns aggregate case and death table with metrics
}
\description{
Load, clean, and analyze case and death tab
}
