% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospital_tbl.R
\name{get_hosp_nat_tbl}
\alias{get_hosp_nat_tbl}
\title{get_hosp_nat_tbl}
\usage{
get_hosp_nat_tbl(
  username,
  date_column = "collection_date",
  lag = 2,
  start_date = "2020-01-01"
)
}
\arguments{
\item{username}{cpu username of user compiling the report.}

\item{date_column}{name of date column to be renamed}

\item{lag}{number of days to remove from the most recent date}

\item{start_date}{first day date to filter}
}
\value{
returns aggregate hospital table with metrics
}
\description{
get_hosp_nat_tbl
}
