% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospital_tbl.R
\name{get_hosp_reg_tbl}
\alias{get_hosp_reg_tbl}
\title{Load, clean, and analyze regional hospital data}
\usage{
get_hosp_reg_tbl(
  username,
  date_column = "collection_date",
  lag = 2,
  start_date = "2020-01-01"
)
}
\arguments{
\item{username}{cpu username of user compiling the report.}

\item{date_column}{name of date column to be renamed}

\item{lag}{number of days to remove from the most recent date}

\item{start_date}{first day date to filter}

\item{state_abbr}{name of column with state and territory abbreviations to be renamed and standardize abbreviations}
}
\value{
returns aggregate hospital table with metrics
}
\description{
Load, clean, and analyze regional hospital data
}
