% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lab_tbl.R
\name{get_lab_nat_tbl}
\alias{get_lab_nat_tbl}
\title{Load, clean, and analyze lab data}
\usage{
get_lab_nat_tbl(
  username,
  date_column = "date",
  state_abbr = "state",
  lag = 3,
  start_date = "2020-03-01"
)
}
\arguments{
\item{username}{cpu username of user compiling the report.}

\item{date_column}{name of date column to be renamed}

\item{state_abbr}{name of column with state and territory abbreviations to be renamed and standardize abbreviations}

\item{lag}{number of days to remove from the most recent date}

\item{start_date}{first day date to filter}
}
\value{
returns aggregate lab table with metrics
}
\description{
Load, clean, and analyze lab data
}
