% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_mean.R
\name{get_mean}
\alias{get_mean}
\title{get_mean}
\usage{
get_mean(column, days = 7)
}
\arguments{
\item{column}{column to calculate mean}

\item{days}{date range, number of days to calculate mean}
}
\value{
mean column
}
\description{
get_mean
}
