% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_metrics.R
\name{get_new_per_100k}
\alias{get_new_per_100k}
\title{get_new_per_100k}
\usage{
get_new_per_100k(column, days = 7, pop = pop)
}
\arguments{
\item{column}{column to calculate mean}

\item{days}{date range, number of days to calculate mean}

\item{pop}{population data}
}
\description{
get_new_per_100k
}
