% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_per_change.R
\name{get_per_change}
\alias{get_per_change}
\title{get_per_change}
\usage{
get_per_change(column, column_date)
}
\arguments{
\item{column}{avg column}

\item{column_date}{date column}
}
\value{
formatted percent change column
}
\description{
get_per_change
}
