% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_list_format.R
\name{numbers_list_format_exsum}
\alias{numbers_list_format_exsum}
\title{Load, clean, and analyze state hospital data}
\usage{
numbers_list_format_exsum(dataset)
}
\arguments{
\item{dataset}{}
}
\value{
returns aggregate hospital table with metrics
}
\description{
Load, clean, and analyze state hospital data
}
