% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_list_format.R
\name{numbers_list_new_format_exsum}
\alias{numbers_list_new_format_exsum}
\title{Title}
\usage{
numbers_list_new_format_exsum(dataset1, dataset2, indicator)
}
\arguments{
\item{dataset2}{}
}
\description{
Title
}
