% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_list_format.R
\name{numbers_list_new_format_exsum2}
\alias{numbers_list_new_format_exsum2}
\title{Title}
\usage{
numbers_list_new_format_exsum2(dataset1, dataset2, indicator)
}
\arguments{
\item{indicator}{}
}
\description{
Title
}
