% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epi_curve.R
\name{plot_epi_curve}
\alias{plot_epi_curve}
\title{plot daily, weekly, and monthly epi curves}
\usage{
plot_epi_curve(data_tbl, timeframe, color)
}
\arguments{
\item{data_tbl}{indicator table with daily values}

\item{timeframe}{indicate timeframe, options are "Daily", "Weekly", and "Monthly"}

\item{color}{color of epicurve bars}
}
\description{
plot daily, weekly, and monthly epi curves
}
