% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epi_curve.R
\name{plot_geofacet_epicurve}
\alias{plot_geofacet_epicurve}
\title{plot geofaceted daily, weekly, and monthly epi curves}
\usage{
plot_geofacet_epicurve(data_tbl, timeframe)
}
\arguments{
\item{data_tbl}{indicator table with daily values}

\item{timeframe}{indicate timeframe, options are "Daily", "Weekly", and "Monthly"}
}
\description{
plot geofaceted daily, weekly, and monthly epi curves
}
