% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_map.R
\name{plot_region_map}
\alias{plot_region_map}
\title{plot_region_map}
\usage{
plot_region_map(region_number = 1)
}
\arguments{
\item{region_number}{Region number}
}
\value{
map of US with region highlighted
}
\description{
plot_region_map
}
