% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ben.R
\name{theme_ben}
\alias{theme_ben}
\title{theme_ben}
\usage{
theme_ben(base_size = 14)
}
\arguments{
\item{base_size}{base font size, given in pts}
}
\value{
Creates a theme which controls all non-data display.
}
\description{
theme_ben
}
